+++ CHANGELOG.md
@@ -29,7 +29,7 @@ If breaking changes are needed do be done, they are:
 
 ### Changes
 
-- `service_identity.(cryptography|pyopenssl).extract_ids()` are now public APIs.
+- `service_identity.(cryptography|pyopenssl).extract_patterns()` are now public APIs (FKA `extract_ids()`).
   You can use them to extract the patterns from a certificate without verifying anything.
   [#55](https://github.com/pyca/service-identity/pull/55)